mikesfunc1/requirements.txt
azure-functions==1.10.1
#azure-functions-worker==1.1.9
#grpcio==1.26.0
#grpcio-tools==1.26.0
#pkg-resources==0.0.0
#protobuf==3.13.0
#six==1.15.0
requests==2.27.1
azure-appconfiguration==1.3.0
